install cronjob + refactor webscrape to func
